Refactor profile and wallet navigation menus
