+++ src/components/Layout/Navbar/WalletNavMenu.tsx
@@ -1,49 +1,62 @@
 import { logEvent } from '@/utils/googleAnalytics'
-import { Icon, UseDisclosureReturn } from '@chakra-ui/react'
+import { Button, UseDisclosureReturn } from '@chakra-ui/react'
 import React from 'react'
-import { RiWalletLine } from 'react-icons/ri'
 import { getUserAddressFromCache } from '@/utils/WalletUtils/getUserAddressFromCache'
+import DisplayAvatar from '@/components/Elements/Avatar/DisplayAvatar'
 
-interface WalletNavMenuProps {
-  setVisibleMenu: React.Dispatch<React.SetStateAction<number | null>>
+export interface WalletNavMenuProps {
+  // setVisibleMenu: React.Dispatch<React.SetStateAction<number | null>>
   setHamburger: React.Dispatch<React.SetStateAction<boolean>>
   drawerOperations: UseDisclosureReturn
 }
 
 const WalletNavMenu = ({
-  setVisibleMenu,
+  // setVisibleMenu,
   setHamburger,
   drawerOperations,
 }: WalletNavMenuProps) => {
+  const userAddress = getUserAddressFromCache()
   const handleWalletIconAction = () => {
     logEvent({
       action: 'OPEN_DRAWER',
       value: 1,
-      label: getUserAddressFromCache() || '',
+      label: userAddress || '',
       category: 'open_drawer',
     })
     setHamburger(false)
     drawerOperations.onToggle()
   }
 
-  return (
-    <Icon
-      display={{
-        base: 'none',
-        md: 'flex',
-      }}
+  return !userAddress ? (
+    <Button
+      size="sm"
+      fontSize="14px"
+      fontWeight={600}
+      onClick={handleWalletIconAction}
+    >
+      Sign In
+    </Button>
+  ) : (
+    <Button
+      variant="unstyled"
       color="linkColor"
       cursor="pointer"
       fontSize="3xl"
       onClick={handleWalletIconAction}
       fontWeight={600}
-      as={RiWalletLine}
-      onMouseEnter={() => setVisibleMenu(null)}
+      // onMouseEnter={() => setVisibleMenu(null)}
       _hover={{
         textDecoration: 'none',
         color: 'linkHoverColor',
       }}
-    />
+    >
+      <DisplayAvatar
+        key={userAddress}
+        address={userAddress}
+        size={30}
+        alt={userAddress}
+      />
+    </Button>
   )
 }
